+++ src/dreamstate/test-utils/registry/mockManager.test.ts
@@ -1,4 +1,6 @@
+import { IScopeContext } from "@/dreamstate/core/scoping/ScopeContext";
 import { mockManager } from "@/dreamstate/test-utils/registry/mockManager";
+import { mockScope } from "@/dreamstate/test-utils/registry/mockScope";
 import { TestManager } from "@/fixtures";
 
 describe("mockManager test util", () => {
@@ -7,5 +9,19 @@ describe("mockManager test util", () => {
 
     expect(manager).toBeInstanceOf(TestManager);
     expect(manager.context.first).toBe("first");
+  });
+
+  it("Should properly inject manager in provided scope in case of supplied parameter", () => {
+    const scope: IScopeContext = mockScope();
+
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_INSTANCES_REGISTRY.size).toBe(0);
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_STATES_REGISTRY.size).toBe(0);
+
+    const manager: TestManager = mockManager(TestManager, undefined, scope);
+
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_INSTANCES_REGISTRY.size).toBe(1);
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_STATES_REGISTRY.size).toBe(1);
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_INSTANCES_REGISTRY.has(TestManager)).toBeTruthy();
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_INSTANCES_REGISTRY.get(TestManager)).toBe(manager);
   });
 });