+++ src/dreamstate/test-utils/registry/mockManager.ts
@@ -7,14 +7,15 @@ import { IContextManagerConstructor, TAnyObject } from "@/dreamstate/types";
  *
  * @param {IContextManagerConstructor} ManagerClass - class reference of context manager that should be created.
  * @param {TAnyObject} initialState - initial state that should be injected in class constructor.
+ * @param {IScopeContext} scope - optional scope context where manager should be mocked,
+ *   in case of undefined new scope is created.
  * @returns InstanceType<IContextManagerConstructor> manager class instance.
  */
 export function mockManager<T extends TAnyObject, S extends TAnyObject, M extends IContextManagerConstructor<T, S>>(
   ManagerClass: M,
-  initialState?: S
+  initialState?: S,
+  scope: IScopeContext = initializeScopeContext()
 ): InstanceType<M> {
-  const scope: IScopeContext = initializeScopeContext();
-
   scope.INTERNAL.registerService(ManagerClass, initialState);
 
   return scope.INTERNAL.REGISTRY.CONTEXT_INSTANCES_REGISTRY.get(ManagerClass) as InstanceType<M>;